add view model and refactor trello into package
